+++ packages/contracts/test/systems/SpawnSystem.t.sol
@@ -15,16 +15,8 @@ contract SpawnSystemTest is PrimodiumTest {
 
     bool spawned = Spawned.get(world, playerEntity);
     assertTrue(spawned, "Player should have spawned");
-    assertEq(
-      HomeAsteroid.get(world, playerEntity),
-      asteroidEntity,
-      "Player should have spawned on their own asteroid"
-    );
-    assertEq(
-      RockType.get(world, asteroidEntity),
-      uint8(ERock.Asteroid),
-      "Asteroid should be a normal asteroid"
-    );
+    assertEq(HomeAsteroid.get(world, playerEntity), asteroidEntity, "Player should have spawned on their own asteroid");
+    assertEq(RockType.get(world, asteroidEntity), ERock.Asteroid, "Asteroid should be a normal asteroid");
 
     assertEq(Level.get(world, playerEntity), 1, "Player should have level 1");
   }
@@ -39,9 +31,7 @@ contract SpawnSystemTest is PrimodiumTest {
     // Asteroid Count is incremented before creation in createAsteroid(), so the asteroid index starts at one.
     // We create ten asteroids consecutively and check if their assigned coordinates match the expected coordinates based on their order of creation.
     for (uint32 i = 1; i <= 10; i++) {
-      address newAddress = address(
-        uint160(uint256(keccak256(abi.encodePacked(i * 12345))))
-      );
+      address newAddress = address(uint160(uint256(keccak256(abi.encodePacked(i * 12345)))));
       bytes32 playerEntity = addressToEntity(newAddress);
       PositionData memory position = LibAsteroid.getUniqueAsteroidPosition(i);
       spawn(newAddress);
